Refactor: Move business logic to VideoStreamBrowsingHelper
Improve code organization by moving complex database logic from view to helper class:

**Helper Class Enhancements:**
- Added get_timeline_window() method for timestamp-based window queries
- Added find_adjacent_records() method for efficient navigation
- Added get_sensor_history_view_data() high-level method that encapsulates all business logic
- Consolidated all SensorHistory/SensorResponse conversion and query logic

**View Simplification:**
- EntityVideoSensorHistoryView.get_main_template_context() now uses single helper call
- Removed 80+ lines of complex database query logic from view
- View focuses on request/response handling as intended
- Eliminated duplicate navigation and timeline logic

**Better Separation of Concerns:**
- Views handle HTTP request/response and validation
- Helper class handles business logic and database queries
- Timeline windowing, pagination metadata, and navigation all encapsulated
- Easier to test business logic independently of Django views

**Code Quality Benefits:**
- Single source of truth for sensor history queries
- Reusable helper methods for future features
- Cleaner, more maintainable view code
- Business logic properly encapsulated and testable
