+++ src/hi/apps/console/video_stream_browsing_helper.py
@@ -91,6 +91,72 @@ class VideoStreamBrowsingHelper:
         return sensor_response
     
     @classmethod
+    def get_timeline_window(cls, sensor: Sensor, center_record: SensorHistory = None, window_size: int = 50):
+        """
+        Get a timeline window of SensorHistory records around a center record.
+        Designed to support future pagination functionality.
+        
+        Args:
+            sensor: Sensor to get records for
+            center_record: Record to center the window around (None for most recent)
+            window_size: Total number of records to include
+            
+        Returns:
+            Tuple of (sensor_responses_list, pagination_metadata)
+        """
+        if center_record is None:
+            # No center record - get most recent records (default behavior)
+            history_records = SensorHistory.objects.filter(
+                sensor=sensor,
+                has_video_stream=True
+            ).order_by('-response_datetime')[:window_size]
+            
+            has_older_records = len(history_records) == window_size
+            has_newer_records = False
+            window_center_timestamp = history_records[0].response_datetime if history_records else None
+        else:
+            # Center window around specific record
+            half_window = window_size // 2
+            
+            # Get records before center (older timestamps)
+            before_records = list(SensorHistory.objects.filter(
+                sensor=sensor,
+                has_video_stream=True,
+                response_datetime__lt=center_record.response_datetime
+            ).order_by('-response_datetime')[:half_window])
+            
+            # Get records after center (newer timestamps) 
+            after_records = list(SensorHistory.objects.filter(
+                sensor=sensor,
+                has_video_stream=True,
+                response_datetime__gt=center_record.response_datetime
+            ).order_by('response_datetime')[:half_window])
+            
+            # Combine all records in chronological order (newest first)
+            history_records = list(reversed(after_records)) + [center_record] + before_records
+            
+            # Pagination metadata for future use
+            has_older_records = len(before_records) == half_window
+            has_newer_records = len(after_records) == half_window
+            window_center_timestamp = center_record.response_datetime
+        
+        # Convert to SensorResponse objects
+        sensor_responses = []
+        for record in history_records:
+            sensor_response = cls.create_sensor_response_with_history_id(record)
+            sensor_responses.append(sensor_response)
+        
+        # Pagination metadata for future pagination feature
+        pagination_metadata = {
+            'has_older_records': has_older_records,
+            'has_newer_records': has_newer_records,
+            'window_center_timestamp': window_center_timestamp,
+            'window_size': window_size,
+        }
+        
+        return sensor_responses, pagination_metadata
+    
+    @classmethod
     def group_responses_by_time(cls, sensor_responses: List[SensorResponse]) -> List[Dict]:
         """
         Group sensor responses by time period for timeline display.
@@ -185,3 +251,112 @@ class VideoStreamBrowsingHelper:
             return (prev_response, next_response)
         except (ValueError, IndexError):
             return (None, None)
+    
+    @classmethod
+    def find_adjacent_records(cls, sensor: Sensor, current_history_id: int) -> tuple:
+        """
+        Find previous and next SensorHistory records for navigation.
+        Uses database queries for efficient navigation.
+        
+        Args:
+            sensor: Sensor to find records for
+            current_history_id: ID of current SensorHistory record
+            
+        Returns:
+            Tuple of (prev_sensor_response, next_sensor_response), either can be None
+        """
+        if not current_history_id:
+            return (None, None)
+        
+        try:
+            current_record = SensorHistory.objects.get(
+                id=current_history_id,
+                sensor=sensor,
+                has_video_stream=True
+            )
+        except SensorHistory.DoesNotExist:
+            return (None, None)
+        
+        # Find previous record (older timestamp)
+        prev_record = SensorHistory.objects.filter(
+            sensor=sensor,
+            has_video_stream=True,
+            response_datetime__lt=current_record.response_datetime
+        ).order_by('-response_datetime').first()
+        
+        prev_sensor_response = None
+        if prev_record:
+            prev_sensor_response = cls.create_sensor_response_with_history_id(prev_record)
+        
+        # Find next record (newer timestamp)
+        next_record = SensorHistory.objects.filter(
+            sensor=sensor,
+            has_video_stream=True,
+            response_datetime__gt=current_record.response_datetime
+        ).order_by('response_datetime').first()
+        
+        next_sensor_response = None
+        if next_record:
+            next_sensor_response = cls.create_sensor_response_with_history_id(next_record)
+        
+        return (prev_sensor_response, next_sensor_response)
+    
+    @classmethod
+    def get_sensor_history_view_data(cls, sensor: Sensor, sensor_history_id: int = None):
+        """
+        Get all data needed for the sensor history view.
+        High-level method that encapsulates the business logic.
+        
+        Args:
+            sensor: Sensor to get data for
+            sensor_history_id: Optional specific record ID to display
+            
+        Returns:
+            Dict with sensor_responses, current_sensor_response, timeline_groups, 
+            pagination_metadata, prev/next responses
+        """
+        # Smart query strategy: get timeline window based on whether we have a specific record
+        if sensor_history_id:
+            # Specific record requested - get timeline window centered around it
+            try:
+                current_history_record = SensorHistory.objects.get(
+                    id=sensor_history_id,
+                    sensor=sensor,
+                    has_video_stream=True
+                )
+                # Get timeline window centered around this record
+                sensor_responses, pagination_metadata = cls.get_timeline_window(
+                    sensor, current_history_record
+                )
+                # Current record is guaranteed to be in the timeline
+                current_sensor_response = next(
+                    (r for r in sensor_responses 
+                     if r.detail_attrs and r.detail_attrs.get('sensor_history_id') == str(sensor_history_id)),
+                    None
+                )
+            except SensorHistory.DoesNotExist:
+                # Record not found - fall back to most recent window
+                sensor_responses, pagination_metadata = cls.get_timeline_window(sensor, None)
+                current_sensor_response = sensor_responses[0] if sensor_responses else None
+        else:
+            # No specific record - get most recent window
+            sensor_responses, pagination_metadata = cls.get_timeline_window(sensor, None)
+            current_sensor_response = sensor_responses[0] if sensor_responses else None
+        
+        # Group sensor responses by time period
+        timeline_groups = cls.group_responses_by_time(sensor_responses)
+        
+        # Find previous and next responses for navigation
+        current_history_id = sensor_history_id if sensor_history_id else None
+        prev_sensor_response, next_sensor_response = cls.find_adjacent_records(
+            sensor, current_history_id
+        )
+        
+        return {
+            'sensor_responses': sensor_responses,
+            'current_sensor_response': current_sensor_response,
+            'timeline_groups': timeline_groups,
+            'pagination_metadata': pagination_metadata,
+            'prev_sensor_response': prev_sensor_response,
+            'next_sensor_response': next_sensor_response,
+        }
